Base

Description

The ABS' AST and the ABS-runtime data-structures

The values of our language

type Ref = Int

We only have one type for our values: the Reference Both objects and futures use this type Ref. Since in this ABS-subset we are dynamically-typed, it is undefined what happens if an object is expected in place of a future, and vice-versa.

type ObjRef = Ref

Objects have type Ref

type FutRef = Ref

Futures have type Ref

type Counter = Ref

An ever-increasing counter to pick new unique references when creating new objects (via new) and futures (via async) (written as type-synonym just for clarity)

The datastructures used at ABS-runtime

data Heap

the simulated Heap datastructure where all program's objects live in. We assume no (automatic) Garbage-Collection of the Heap for the moment. In reality, no GC will be performed from the underlying Haskell-GC, because we do not delete objects

Constructors

Heap

Fields

objects :: Objects

the live objects

futures :: Futures

the live futures

newRef :: Counter

the counter

type Objects = Map ObjRef Attrs

The objects of the heap is a table _from_ an object's Reference _to_ another table of the object's attributes: Attrs

type Attrs = Map String Ref

The attributes is a table _from_ the attribute name (String) _to_ its value (Ref)

type Futures = Map FutRef (Maybe Ref)

The futures of the heap is a table _of_ future's References _to_ their potential final values. A future is empty (Nothing) until it is resolved (filled with 'Just value'), hence its type 'Maybe Ref'. A future reference must be final, and this must be guaranteed by the runtime system.

type Cont = () -> Stmt

We have a single (universal) type for our continuations. Later, if we introduce local-variables we are going to need an extra type for Continuations: 'Ref -> Stmt'

newtype Proc

Each process is a triple of the this object, its destiny, and its (resumable) continuation (note: is a newtype just for overriding its Show instance, check module PP)

Constructors

Proc

Fields

fromProc :: (ObjRef, FutRef, Cont)

type ProcTable = Map ObjRef [Proc]

The (global) scheduler's runtime Process Table. It is a mapping _from_ an object-reference _to_ its queue of processes

Our language's AST and types

data Stmt

An ABS statement. Statements are "chained" (sequantially composed) by deeply nesting them through Continuations.

Constructors

Assign String Rhs Cont
"attr" := Rhs; cont...

Await String Cont
await "attr"; cont...

If BExp (Cont -> Stmt) (Cont -> Stmt) Cont
if pred ThenClause ElseClause; cont...

While BExp (Cont -> Stmt) Cont
while pred BodyClause; cont...

Skip Cont
skip; cont...

Return String (Maybe String) Cont
return "attr" WriteBack; cont... (note: if it is a sync call then we pass as an argument to return, the attribute to write back to, if it is async call then we pass Nothing)

Stop
is not used explictly by any ABS-program; it is used by the runtime to signal the end of a chain of statements; thus, it does not take a continuation as an argument.

data Rhs

the RHS of an assignment

Constructors

New

Get String

Async String Method [String]

Sync Method [String]

Param Ref
all commands operate on attributes; to use instead a method's parameter (passed argument or this) you first store it to an (auxiliary) attribute, e.g. Assign "attr" (Param this)

Attr String
assign an attribute to the value of another attribute we do not need This, because it is passed as a local parameter on each method

data BExp

A boolean expression occurs only as a control-flow predicate (if & while) It does only reference equality of attributes (BEq) and combinators on them (conjuction,disjunction,negation).

Constructors

BEq String String

BNeg BExp

BCon BExp BExp

BDis BExp BExp

type Method

Arguments

= [Ref]
a list of passed (deref) parameters

-> ObjRef
this obj

-> Maybe String
in case of sync call: a writeback attribute to write the return result to

-> Cont
the continuation after the method is finished

-> Cont
the resulting method's continuation that will start executing when applied to ()

The type of every top-level ABS-method.