Eval

Description

Contains the single-step evaluator/interpreter of ABS terms (from AST)

Documentation

eval

Arguments

:: Proc
a process to execute

-> Heap
inside a heap

-> (Stmt, Maybe Cont, Maybe Proc, Heap)

eval takes a process and a heap and executes the 1st only statement of that process. Returns the executed statement, maybe the rest continuation of this process, maybe a brand-new process (from async) and a new heap.

beval :: BExp -> ObjRef -> Heap -> Bool

Evaluates a predicate BExp from the AST to a Haskell's Bool

readAttr

Arguments

:: String
the name of the attribute

-> ObjRef
the this object

-> Heap
the current heap

-> Ref
its value

Utility function to read an attribute from an object (this)