Sched

Description

The global scheduler that schedules COGs (single-objects in our case) We simulate non-determinism by pseudo-randomness

Documentation

run :: Int -> Method -> IO Heap

Given a number of iterations and an entrypoint (ABS main-method) , it executes the program only for that many iterations returning the final heap.

sched

Arguments

:: Int
the max iterations to run

-> (Heap, SchedQueue)
an initial program configuration

-> IO Heap
the final heap

The sched is just a looper of the step function