Sched

Description

The global scheduler that schedules COGs (single-objects in our case) We simulate non-determinism by pseudo-randomness

Documentation

run :: Int -> Method -> StdGen -> Heap

Given a number of iterations and an entrypoint (ABS main-method) , it executes the program only for that many iterations returning the final heap.

runFixed :: Int -> Method -> Heap

Wrapper to run using a default (fixed) stdgen

runIO :: Int -> Method -> IO Heap

Wrapper to run using a "true" stdgen from the hardware clock

sched

Arguments

:: Int
the max iterations to run

-> (Heap, ProcTable, StdGen)
an initial program configuration

-> Heap
the final heap

The sched is just a looper of the step function

step

Arguments

:: (Heap, ProcTable, StdGen)
current configuration

-> Maybe (Heap, ProcTable, StdGen)
new configuration

The step is the main workforce. Given a current configuration, it randomly picks up a next object to execute, then pops out its first process, evaluates the 1st statement of that process and returns either a modified next configuration or Nothing to signal that there was no object left to execute (program finished successfully). Note: Favourably, the step should check if there is no _active_ object left (program finished or program deadlocked).